% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstantMatrix.R
\name{ConstantMatrix}
\alias{ConstantMatrix}
\alias{ConstantMatrix-class}
\alias{ConstantMatrixSeed-class}
\alias{extract_array,ConstantMatrixSeed-method}
\alias{DelayedArray,ConstantMatrixSeed-method}
\alias{ConstantMatrixSeed}
\title{A DelayedMatrix containing a constant value}
\usage{
\S4method{DelayedArray}{ConstantMatrixSeed}(seed)

ConstantMatrixSeed(dim, value)

ConstantMatrix(dim, value)
}
\arguments{
\item{seed}{A ConstantMatrixSeed object.}

\item{dim}{Integer vector of length 2, containing the dimensions of the matrix.}

\item{value}{Atomic scalar containing the value to fill the matrix.}
}
\value{
The \code{ConstantMatrixSeed} constructor returns a ConstantMatrixSeed object.

The \code{ConstantMatrix} and \code{DelayedArray} constructors return a ConstantMatrix object.
}
\description{
A DelayedMatrix backend to efficiently mimic a matrix containing a constant value, without actually creating said matrix in memory.
}
\details{
This class allows us to efficiently create placeholder matrices, primarily to be used in \code{\link{unsplitAltExps}}.
For example, we can create a matrix full of \code{NA} values to enable combining of matrices without counterparts in other alternative experiments.

We create this class instead of using the \linkS4class{RleArray},
as the latter requires some workarounds when the product of the dimensions is greater than the maximum integer value.
}
\examples{
# This would ordinarily take up 8 TB of memory:
out <- ConstantMatrix(c(1e6, 1e6), value=NA_real_)
out

}
\author{
Aaron Lun
}
