% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEInput.R
\docType{class}
\name{SCEInput}
\alias{SCEInput}
\alias{MainExpInput}
\alias{AssayInput}
\alias{ReducedDimInput}
\alias{AltExpInput}
\alias{AltAssayInput}
\alias{AltReducedDimInput}
\alias{SCEInput-class}
\alias{MainExpInput-class}
\alias{AssayInput-class}
\alias{ReducedDimInput-class}
\alias{AltExpInput-class}
\alias{AltAssayInput-class}
\alias{AltReducedDimInput-class}
\alias{show}
\alias{show,SCEInput-method}
\alias{getInput}
\alias{getInput,SingleCellExperiment,MainExpInput-method}
\alias{getInput,SingleCellExperiment,AssayInput-method}
\alias{getInput,SingleCellExperiment,ReducedDimInput-method}
\alias{getInput,SingleCellExperiment,AltExpInput-method}
\alias{getInput,SingleCellExperiment,AltAssayInput-method}
\alias{getInput,SingleCellExperiment,AltReducedDimInput-method}
\alias{isInputReducedDim}
\alias{isInputReducedDim,SCEInput-method}
\alias{isInputReducedDim,ReducedDimInput-method}
\alias{isInputReducedDim,AltReducedDimInput-method}
\alias{isInputSameShape}
\alias{isInputSameShape,SCEInput-method}
\alias{isInputSameShape,ReducedDimInput-method}
\alias{isInputSameShape,AltReducedDimInput-method}
\alias{isInputAltExp}
\alias{isInputAltExp,SCEInput-method}
\alias{isInputAltExp,AltExpInput-method}
\alias{inputArguments}
\alias{makeAllExpInputs}
\alias{makeSameAssayInputs}
\title{The SCEInput class}
\description{
The SCEInput classes provide a flexible approach to extracting different types of data from a \linkS4class{SingleCellExperiment} object.
Each subclass specifies a different piece of data to fetch from the SingleCellExperiment in a generic manner,
as well as storing custom arguments that can be to process that particular piece of data.
This is helpful for general-purpose functions like \code{\link{applySCE}},
as well as complex functions that use multiple pieces of information from a SingleCellExperiment object.
}
\section{Constructors}{

One constructor is available for each SCEInput subclass:
\itemize{
\item \code{MainExpInput(...)}
\item \code{AssayInput(assay, ...)}
\item \code{ReducedDimInput(type, ...)}
\item \code{AltExpInput(experiment, ...)}
\item \code{AltAssayInput(experiment, assay, ...)}
\item \code{AltReducedDimInput(experiment, type, ...)}
}

With the following arguments:
\describe{
\item{...}{Aguments (usually named) to pass to other functions, e.g., \code{FUN} in \code{\link{applySCE}}.}
\item{assay}{String or integer specifying the assay to use.}
\item{type}{String or integer specifying the \code{\link{reducedDim}} to use.}
\item{experiment}{String or integer specifying the \code{\link{altExp}} to use.}
}
}

\section{Getting the data}{

\code{getInput(x, input)} will return the specified piece of data from a \linkS4class{SingleCellExperiment} \code{x},
based on the SCEInput object \code{input}.
\itemize{
\item If \code{input} is a MainExpInput, \code{x} is returned directly.
\item If \code{input} is a AssayInput, the specified assay of \code{x} is returned. 
\item If \code{input} is a ReducedDimInput, the specified entry of \code{\link{reducedDims}(x)} is returned. 
\item If \code{input} is an AltExpInput, the specified alternative Experiment in \code{x} is returned.
\item If \code{input} is an AltAssayInput, the specified assay of the specified alternativeExperiment is returned.
\item If \code{input} is an AltReducedDimInput, the specified \code{\link{reducedDims}} of the specified alternativeExperiment is returned.
}
}

\section{Input properties}{

In the following code snippets, \code{input} is a SCEInput object.

\code{inputArguments(input)} will return a list of arguments from \code{input}.

\code{isReducedDimInput(input)} will return \code{TRUE} for ReducedDimInput and AltExpReducedInput objects, and \code{FALSE} otherwise.

\code{isInputSameShape(input)} will return \code{TRUE} for all \code{input} where \code{getInput(x, input)} is guaranteed to have the same dimensions as \code{x}.

\code{isInputAltExp(input)} will return \code{TRUE} for all \code{input} where \code{getInput(x, input)} pulls data from \code{altExps(x)}.
}

\section{Constructing lists of SCEInputs}{

In the following code snippets, \code{x} is a SingleCellExperiment object.

\code{makeAllExpInputs(x, include.main=TRUE)} will return a list containing AltExpInputs for all \code{\link{altExpNames}(x)}.
If \code{include.main=TRUE}, a MainExpInput is prepended to the list.

\code{makeSameAssayInputs(x, assay, include.main=TRUE)} will return a list containing AltAssayInputs with the specified \code{assay} for all \code{\link{altExpNames}(x)}.
If \code{include.main=TRUE}, an AssayInput is prepended to the list.
}

\examples{
ncells <- 100
u <- matrix(rpois(20000, 5), ncol=ncells)
pca <- matrix(runif(ncells*5), ncells)
sce <- SingleCellExperiment(assays=list(counts=u),
    reducedDims=SimpleList(PCA=pca))
sce.copy <- SingleCellExperiment(assays=list(counts=u*10),
    reducedDims=SimpleList(PCA=pca/10))
altExp(sce, "BLAH") <- sce.copy

getInput(sce, MainExpInput())
str(getInput(sce, AssayInput("counts")))
str(getInput(sce, ReducedDimInput("PCA")))
getInput(sce, AltExpInput("BLAH"))
str(getInput(sce, AltAssayInput("BLAH", "counts")))
str(getInput(sce, AltReducedDimInput("BLAH", "PCA")))

}
\seealso{
\code{\link{applySCE}}, for an example of how this can be used in a generic manner.
}
\author{
Aaron Lun
}
