% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyToSCE.R
\name{simplifyToSCE}
\alias{simplifyToSCE}
\title{Simplify a list to a single SingleCellExperiment}
\usage{
simplifyToSCE(results, which, x)
}
\arguments{
\item{results}{A list of SummarizedExperiment or SingleCellExperiment objects.}

\item{which}{A list of \linkS4class{SCEInput} objects of the same length as \code{results}.
This should be of the same length as that used to construct \code{results}.}

\item{x}{The original SingleCellExperiment from which \code{results} were computed.}
}
\value{
A SingleCellExperiment corresponding to the entry of \code{results} generated from any part of the main Experiment of \code{x}.
All results generated from the alternative Experiments of \code{x} are stored in the \code{\link{altExps}} of the output.

Alternatively, if simplification could not be performed, \code{NULL} is returned with a warning.
}
\description{
Simplify a list of \linkS4class{SingleCellExperiment}, usually generated by \code{\link{applySCE}} on main and alternative Experiments,
into a single SingleCellExperiment containing some of the results in its \code{\link{altExps}}.
}
\details{
Each entry of \code{results} should have the same number and names of the columns.

In \code{which}, there should be no more than one entry that extracts data from the main Experiment 
(e.g., a \linkS4class{MainExpInput}, \linkS4class{AssayInput} or \linkS4class{ReducedDimInput}).
If one exists, the corresponding entry of \code{results} should be a SingleCellExperiment and is then used as the main Experiment in the output.
Otherwise, if no such entry of \code{which} exists, an empty SingleCellExperiment is constructed as a container for the \code{\link{altExps}}.

In \code{which}, there should be no more than one reference to each alternative Experiment in \code{x}.
For each reference, the corresponding entry of \code{results} is stored in the \code{\link{altExps}} of the output under the same name that was used in \code{x}.
}
\examples{
ncells <- 100
u <- matrix(rpois(20000, 5), ncol=ncells)
pca <- matrix(runif(ncells*5), ncells)
sce <- SingleCellExperiment(assays=list(counts=u),
    reducedDims=SimpleList(PCA=pca))
altExp(sce, "BLAH") <- SingleCellExperiment(assays=list(counts=u*10),
    reducedDims=SimpleList(PCA=pca/10))
altExp(sce, "WHEE") <- SingleCellExperiment(assays=list(counts=u*2),
    reducedDims=SimpleList(PCA=pca/5))

# Setting FUN=identity just extracts each piece:
which <- list(MainExpInput(), AltExpInput("BLAH"), AltExpInput("WHEE"))
results <- applySCE(sce, FUN=identity, SIMPLIFY=FALSE, which=which)
results

# Simplifying to an output that mirrors the structure of 'sce'.
simplifyToSCE(results, which, sce)

}
\seealso{
\code{\link{applySCE}}, where this function is used when \code{SIMPLIFY=TRUE}.
}
\author{
Aaron Lun
}
